requirements.txt
pyspark==3.3.1
delta-spark==2.1.0